Renamed game-end-type file
