+++ src/common/game-end-reasons.ts
